packed storage compatability
adds the ability to craft wooden hoppers using packed storage's chests, since the mod makes the vanilla chest uncraftable.
this will make any type of packed chest work for any of the recipes, with the outcome depending on the planks used. you may want to fix that, but for myself this is good enough.

im not familiar with modding, so i dont know if this will cause a crash or any problems if used without packed storage.